Add menucreator
Added a menucreator to create a menu out of an array. This array contains menuitems which contain at least two things, a title and link to that page. As a third option the menuitem can contain an items variable which can contain an entire menu array of itself to make dropdown menus. This third option is optionable and is not required in the array, if not existent there will simply be no dropdown menu.